prototype: testing environment for gm fitting. i realised that negative bias is problematic (see description)
if the bias is negative, then the resulting function will have infinite
support. it'd be therefore impossible to fit a gaussian mixture (without
adding a constant). I believe a constant wouldn't hurt in the
convolution itself (it transforms to a dirac at 0/0, and then it can be
multiplied, no?). but we have a constant and clamp the function. the
resulting beast (`¬-"\____/`^-*, see the flat line on the bottom) would
be even harder to fit. i believe it would produce a special case (at
least), plus it's hard to model a flat surface with gaussians.